Organise the classes into suitable java packages
The code has been organised into suitable java packages as followed:

* An overall duke package
* Packages in duke: command, exception, storage, task, ui
