--- src/main/java/Duke.java
@@ -1,167 +1,0 @@
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
-
-class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    public void markAsNotDone() {
-        isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "[X]" : "[ ]");
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    @Override
-    public String toString() {
-        return this.getStatusIcon() + " " + this.getDescription();
-    }
-
-    public String toFormattedString() {
-        return "";
-    }
-
-    public static Task createTaskFromFormattedString(String formattedString) {
-        String[] parts = formattedString.split(" \\| ");
-        String taskType = parts[0];
-        boolean isDone = parts[1].equals("1");
-        String description = parts[2];
-
-        if (taskType.equals("T")) {
-            Task todo =  new ToDo(description);
-            todo.isDone = isDone;
-            return todo;
-        } else if (taskType.equals("D")) {
-            String by = parts[3];
-            Task deadline =  new Deadline(description, by);
-            deadline.isDone = isDone;
-            return deadline;
-        } else if (taskType.equals("E")) {
-            String from = parts[3];
-            String to = parts[4];
-            Task event = new Event(description, from, to);
-            event.isDone = isDone;
-            return event;
-        } else {
-            // Handle unrecognized task type
-            return null;
-        }
-    }
-}
-
-class ToDo extends Task {
-    public ToDo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    @Override
-    public String toFormattedString() {
-        return "T | " + (isDone ? "1" : "0") + " | " + description;
-    }
-}
-
-class Deadline extends Task {
-    protected LocalDateTime by;
-    public Deadline(String description, String by) {
-        super(description);
-        this.by = LocalDateTime.parse(by);
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")) + ")";
-    }
-
-    @Override
-    public String toFormattedString() {
-        return "D | " + (isDone ? "1" : "0") + " | " + description + " | " + by;
-    }
-}
-
-class Event extends Task {
-    protected LocalDateTime from;
-    protected LocalDateTime to;
-
-    public Event(String description, String from, String to){
-        super(description);
-        this.from = LocalDateTime.parse(from);
-        this.to = LocalDateTime.parse(to);
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + from.format(DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm"))
-                + " to: " + to.format(DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")) + ")";
-    }
-
-    @Override
-    public String toFormattedString() {
-        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + from + " | " + to;
-    }
-}
-
-public class Duke {
-
-    private final Storage storage;
-    private TaskList tasks;
-    private final Ui ui;
-
-    public Duke(String filePath) {
-        ui = new Ui();
-        storage = new Storage(filePath);
-        try {
-            tasks = new TaskList(storage.load());
-        } catch (DukeException e) {
-            ui.showLoadingError();
-            tasks = new TaskList();
-        }
-    }
-
-    public void run() {
-        ui.showWelcome();
-        boolean isExit = false;
-        while (!isExit) {
-            try {
-                String fullCommand = ui.readCommand();
-                ui.showLine(); // show the divider line ("_______")
-                Command c = Parser.parse(fullCommand);
-                c.execute(tasks, ui, storage);
-                isExit = c.isExit();
-            } catch (DukeException e) {
-                ui.showError(e.getMessage());
-            } finally {
-                ui.showLine();
-            }
-        }
-    }
-
-    public static void main(String[] args) {
-        new Duke("./src/main/java/duke.txt").run();
-    }
-}
-
