+++ src/main/java/duke/command/AddTodoCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.task.ToDo;
+import duke.ui.Ui;
+
 public class AddTodoCommand extends Command {
     private final String description;
