+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,15 +1,23 @@
-public class MarkAsUndoneCommand extends Command {
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class DeleteCommand extends Command {
     private final int taskIndex;
 
-    public MarkAsUndoneCommand(int taskIndex) {
+    public DeleteCommand(int taskIndex) {
         this.taskIndex = taskIndex - 1;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         if (taskIndex >= 0 && taskIndex < tasks.size()) {
-            tasks.markTaskAsUndone(taskIndex);
-            ui.showTaskMarkedAsUndone(tasks, taskIndex);
+            Task removedTask = tasks.deleteTask(taskIndex);
+            ui.showTaskRemoved(tasks, removedTask);
         } else {
             throw new DukeException("Invalid task number.");
         }
