+++ src/main/java/duke/command/MarkAsUndoneCommand.java
@@ -1,15 +1,22 @@
-public class MarkAsDoneCommand extends Command {
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class MarkAsUndoneCommand extends Command {
     private final int taskIndex;
 
-    public MarkAsDoneCommand(int taskIndex) {
+    public MarkAsUndoneCommand(int taskIndex) {
         this.taskIndex = taskIndex - 1;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         if (taskIndex >= 0 && taskIndex < tasks.size()) {
-            tasks.markTaskAsDone(taskIndex);
-            ui.showTaskMarkedAsDone(tasks, taskIndex);
+            tasks.markTaskAsUndone(taskIndex);
+            ui.showTaskMarkedAsUndone(tasks, taskIndex);
         } else {
             throw new DukeException("Invalid task number.");
         }
