+++ src/main/java/duke/task/Deadline.java
@@ -1,0 +1,22 @@
+package duke.task;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
+    protected LocalDateTime by;
+    public Deadline(String description, String by) {
+        super(description);
+        this.by = LocalDateTime.parse(by);
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm")) + ")";
+    }
+
+    @Override
+    public String toFormattedString() {
+        return "D | " + (isDone ? "1" : "0") + " | " + description + " | " + by;
+    }
+}
