+++ src/main/java/duke/ui/Ui.java
@@ -1,3 +1,8 @@
+package duke.ui;
+
+import duke.task.TaskList;
+import duke.task.Task;
+
 import java.util.Scanner;
 
 public class Ui {
